Update for new kernel rules
