+++ .eslintrc.json
@@ -19,21 +19,23 @@
         "semi-spacing": [ "error", {"before": false, "after": true} ],
         "comma-spacing": [ "error", { "before": false, "after": true } ],
         "comma-dangle": ["error", "never"],
-        "camelcase": "error",
         "eqeqeq": "error"
     },
     "globals" : {
-        "_elm_community$webgl$Native_Texture": true,
-        "_elm_community$webgl$Native_WebGL": true,
-        "_elm_community$webgl$Native_Settings": true,
-        "_elm_lang$virtual_dom$Native_VirtualDom": false,
-        "_elm_lang$core$Native_Utils": false,
-        "_elm_lang$core$Native_Scheduler": false,
+        "__Texture_SizeError": false,
+        "__Texture_LoadError": false,
+        "__Utils_Tuple2": false,
+        "_VirtualDom_doc": false,
+        "_VirtualDom_custom": false,
+        "__Scheduler_binding": false,
+        "__Scheduler_succeed": false,
+        "__Scheduler_fail": false,
         "F2": false,
         "F3": false,
         "F4": false,
         "F5": false,
         "F6": false,
+        "A2": false,
         "Float32Array": false,
         "Int32Array": false,
         "Uint16Array": false
