+++ elm.json
@@ -1,6 +1,6 @@
 {
     "type": "package",
-    "name": "elm-community/webgl",
+    "name": "elm-explorations/webgl",
     "summary": "A library for general 3D rendering with WebGL",
     "license": "BSD-3-Clause",
     "version": "2.0.5",
