+++ src/Elm/Kernel/WebGL.js
@@ -1,0 +1,622 @@
+/*
+
+import Elm.Kernel.VirtualDom exposing (custom, doc)
+
+*/
+
+function _WebGL_log(/* msg */) {
+  // console.log(msg);
+}
+
+var _WebGL_guid = 0;
+
+function _WebGL_listEach(fn, list) {
+  for (; list.b; list = list.b) {
+    fn(list.a);
+  }
+}
+
+function _WebGL_listLength(list) {
+  var length = 0;
+  for (; list.b; list = list.b) {
+    length++;
+  }
+  return length;
+}
+
+var _WebGL_rAF = typeof requestAnimationFrame !== 'undefined' ?
+  requestAnimationFrame :
+  function (cb) { setTimeout(cb, 1000 / 60); };
+
+// eslint-disable-next-line no-unused-vars
+var _WebGL_unsafeCoerceGLSL = function (src) {
+  return { src: src };
+};
+
+// eslint-disable-next-line no-unused-vars
+var _WebGL_entity = F5(function (settings, vert, frag, mesh, uniforms) {
+
+  if (!mesh.__id) {
+    mesh.__id = _WebGL_guid++;
+  }
+
+  return {
+    // TODO: $ for Entity?
+    $: 'Entity',
+    __vert: vert,
+    __frag: frag,
+    __mesh: mesh,
+    __uniforms: uniforms,
+    __settings: settings
+  };
+
+});
+
+/**
+ *  Apply setting to the gl context
+ *
+ *  @param {WebGLRenderingContext} gl context
+ *  @param {Setting} setting coming in from Elm
+ */
+function _WebGL_applySetting(gl, setting) {
+  // TODO: how to check $?
+  switch (setting.$) {
+    case 'Blend':
+      gl.enable(gl.BLEND);
+      // eq1 f11 f12 eq2 f21 f22 r g b a
+      gl.blendEquationSeparate(setting.a, setting.d);
+      gl.blendFuncSeparate(setting.b, setting.c, setting.e, setting.f);
+      gl.blendColor(setting.g, setting.h, setting.i, setting.j);
+      break;
+    case 'DepthTest':
+      gl.enable(gl.DEPTH_TEST);
+      // func mask near far
+      gl.depthFunc(setting.a);
+      gl.depthMask(setting.b);
+      gl.depthRange(setting.c, setting.d);
+      break;
+    case 'StencilTest':
+      gl.enable(gl.STENCIL_TEST);
+      // ref mask writeMask test1 fail1 zfail1 zpass1 test2 fail2 zfail2 zpass2
+      gl.stencilFuncSeparate(gl.FRONT, setting.d, setting.a, setting.b);
+      gl.stencilOpSeparate(gl.FRONT, setting.e, setting.f, setting.g);
+      gl.stencilMaskSeparate(gl.FRONT, setting.c);
+      gl.stencilFuncSeparate(gl.BACK, setting.h, setting.a, setting.b);
+      gl.stencilOpSeparate(gl.BACK, setting.i, setting.j, setting.b0);
+      gl.stencilMaskSeparate(gl.BACK, setting.c);
+      break;
+    case 'Scissor':
+      gl.enable(gl.SCISSOR_TEST);
+      gl.scissor(setting.a, setting.b, setting.c, setting.d);
+      break;
+    case 'ColorMask':
+      gl.colorMask(setting.a, setting.b, setting.c, setting.d);
+      break;
+    case 'CullFace':
+      gl.enable(gl.CULL_FACE);
+      gl.cullFace(setting.a);
+      break;
+    case 'PolygonOffset':
+      gl.enable(gl.POLYGON_OFFSET_FILL);
+      gl.polygonOffset(setting.a, setting.b);
+      break;
+    case 'SampleCoverage':
+      gl.enable(gl.SAMPLE_COVERAGE);
+      gl.sampleCoverage(setting.a, setting.b);
+      break;
+    case 'SampleAlphaToCoverage':
+      gl.enable(gl.SAMPLE_ALPHA_TO_COVERAGE);
+      break;
+  }
+}
+
+/**
+ *  Revert setting that was applied to the gl context
+ *
+ *  @param {WebGLRenderingContext} gl context
+ *  @param {Setting} setting coming in from Elm
+ */
+function _WebGL_revertSetting(gl, setting) {
+  // TODO: how to check $?
+  switch (setting.$) {
+    case 'Blend':
+      gl.disable(gl.BLEND);
+      break;
+    case 'DepthTest':
+      gl.disable(gl.DEPTH_TEST);
+      break;
+    case 'StencilTest':
+      gl.disable(gl.STENCIL_TEST);
+      break;
+    case 'Scissor':
+      gl.disable(gl.SCISSOR_TEST);
+      break;
+    case 'ColorMask':
+      gl.colorMask(true, true, true, true);
+      break;
+    case 'CullFace':
+      gl.disable(gl.CULL_FACE);
+      break;
+    case 'PolygonOffset':
+      gl.disable(gl.POLYGON_OFFSET_FILL);
+      break;
+    case 'SampleCoverage':
+      gl.disable(gl.SAMPLE_COVERAGE);
+      break;
+    case 'SampleAlphaToCoverage':
+      gl.disable(gl.SAMPLE_ALPHA_TO_COVERAGE);
+      break;
+  }
+}
+
+function _WebGL_doCompile(gl, src, type) {
+
+  var shader = gl.createShader(type);
+  _WebGL_log('Created shader');
+
+  gl.shaderSource(shader, src);
+  gl.compileShader(shader);
+  if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
+    throw gl.getShaderInfoLog(shader);
+  }
+
+  return shader;
+
+}
+
+function _WebGL_doLink(gl, vshader, fshader) {
+
+  var program = gl.createProgram();
+  _WebGL_log('Created program');
+
+  gl.attachShader(program, vshader);
+  gl.attachShader(program, fshader);
+  gl.linkProgram(program);
+  if (!gl.getProgramParameter(program, gl.LINK_STATUS)) {
+    throw gl.getProgramInfoLog(program);
+  }
+
+  return program;
+
+}
+
+function _WebGL_getAttributeInfo(gl, type) {
+  switch (type) {
+    case gl.FLOAT:
+      return { size: 1, type: Float32Array, baseType: gl.FLOAT };
+    case gl.FLOAT_VEC2:
+      return { size: 2, type: Float32Array, baseType: gl.FLOAT };
+    case gl.FLOAT_VEC3:
+      return { size: 3, type: Float32Array, baseType: gl.FLOAT };
+    case gl.FLOAT_VEC4:
+      return { size: 4, type: Float32Array, baseType: gl.FLOAT };
+    case gl.INT:
+      return { size: 1, type: Int32Array, baseType: gl.INT };
+    case gl.INT_VEC2:
+      return { size: 2, type: Int32Array, baseType: gl.INT };
+    case gl.INT_VEC3:
+      return { size: 3, type: Int32Array, baseType: gl.INT };
+    case gl.INT_VEC4:
+      return { size: 4, type: Int32Array, baseType: gl.INT };
+  }
+}
+
+/**
+ *  Form the buffer for a given attribute.
+ *
+ *  @param {WebGLRenderingContext} gl context
+ *  @param {WebGLActiveInfo} attribute the attribute to bind to.
+ *         We use its name to grab the record by name and also to know
+ *         how many elements we need to grab.
+ *  @param {Mesh} mesh The mesh coming in from Elm.
+ *  @return {WebGLBuffer}
+ */
+function _WebGL_doBindAttribute(gl, attribute, mesh) {
+  // The length of the number of vertices that
+  // complete one 'thing' based on the drawing mode.
+  // ie, 2 for Lines, 3 for Triangles, etc.
+  var elemSize = mesh.a.elemSize;
+
+  var idxKeys = [];
+  for (var i = 0; i < elemSize; i++) {
+    idxKeys.push(String.fromCharCode(97 + i));
+  }
+
+  function dataFill(data, cnt, fillOffset, elem, key) {
+    var i;
+    if (elemSize === 1) {
+      for (i = 0; i < cnt; i++) {
+        data[fillOffset++] = cnt === 1 ? elem[key] : elem[key][i];
+      }
+    } else {
+      idxKeys.forEach(function (idx) {
+        for (i = 0; i < cnt; i++) {
+          data[fillOffset++] = cnt === 1 ? elem[idx][key] : elem[idx][key][i];
+        }
+      });
+    }
+  }
+
+  var attributeInfo = _WebGL_getAttributeInfo(gl, attribute.type);
+
+  if (attributeInfo === undefined) {
+    throw new Error('No info available for: ' + attribute.type);
+  }
+
+  var dataIdx = 0;
+  var array = new attributeInfo.type(_WebGL_listLength(mesh.b) * attributeInfo.size * elemSize);
+
+  _WebGL_listEach(function (elem) {
+    dataFill(array, attributeInfo.size, dataIdx, elem, attribute.name);
+    dataIdx += attributeInfo.size * elemSize;
+  }, mesh.b);
+
+  var buffer = gl.createBuffer();
+  _WebGL_log('Created attribute buffer ' + attribute.name);
+
+  gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
+  gl.bufferData(gl.ARRAY_BUFFER, array, gl.STATIC_DRAW);
+  return buffer;
+}
+
+/**
+ *  This sets up the binding caching buffers.
+ *
+ *  We don't actually bind any buffers now except for the indices buffer.
+ *  The problem with filling the buffers here is that it is possible to
+ *  have a buffer shared between two webgl shaders;
+ *  which could have different active attributes. If we bind it here against
+ *  a particular program, we might not bind them all. That final bind is now
+ *  done right before drawing.
+ *
+ *  @param {WebGLRenderingContext} gl context
+ *  @param {Mesh} mesh a mesh object from Elm
+ *  @return {Object} buffer - an object with the following properties
+ *  @return {Number} buffer.numIndices
+ *  @return {WebGLBuffer} buffer.indexBuffer
+ *  @return {Object} buffer.buffers - will be used to buffer attributes
+ */
+function _WebGL_doBindSetup(gl, mesh) {
+  _WebGL_log('Created index buffer');
+  var indexBuffer = gl.createBuffer();
+  var indices = (mesh.a.indexSize === 0)
+    ? _WebGL_makeSequentialBuffer(mesh.a.elemSize * _WebGL_listLength(mesh.b))
+    : _WebGL_makeIndexedBuffer(mesh.c, mesh.a.indexSize);
+
+  gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, indexBuffer);
+  gl.bufferData(gl.ELEMENT_ARRAY_BUFFER, indices, gl.STATIC_DRAW);
+
+  return {
+    numIndices: indices.length,
+    indexBuffer: indexBuffer,
+    buffers: {}
+  };
+}
+
+/**
+ *  Create an indices array and fill it with 0..n
+ *
+ *  @param {Number} numIndices The number of indices
+ *  @return {Uint16Array} indices
+ */
+function _WebGL_makeSequentialBuffer(numIndices) {
+  var indices = new Uint16Array(numIndices);
+  for (var i = 0; i < numIndices; i++) {
+    indices[i] = i;
+  }
+  return indices;
+}
+
+/**
+ *  Create an indices array and fill it from indices
+ *  based on the size of the index
+ *
+ *  @param {List} indicesList the list of indices
+ *  @param {Number} indexSize the size of the index
+ *  @return {Uint16Array} indices
+ */
+function _WebGL_makeIndexedBuffer(indicesList, indexSize) {
+  var indices = new Uint16Array(_WebGL_listLength(indicesList) * indexSize);
+  var fillOffset = 0;
+  var i;
+  _WebGL_listEach(function (elem) {
+    if (indexSize === 1) {
+      indices[fillOffset++] = elem;
+    } else {
+      for (i = 0; i < indexSize; i++) {
+        indices[fillOffset++] = elem[String.fromCharCode(97 + i)];
+      }
+    }
+  }, indicesList);
+  return indices;
+}
+
+function _WebGL_getProgID(vertID, fragID) {
+  return vertID + '#' + fragID;
+}
+
+var _WebGL_drawGL = F2(function (model, domNode) {
+
+  var gl = model.__cache.gl;
+
+  if (!gl) {
+    return domNode;
+  }
+
+  gl.viewport(0, 0, gl.drawingBufferWidth, gl.drawingBufferHeight);
+  gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT | gl.STENCIL_BUFFER_BIT);
+  _WebGL_log('Drawing');
+
+  function drawEntity(entity) {
+    if (!entity.__mesh.b.b) {
+      return; // Empty list
+    }
+
+    var progid;
+    var program;
+    if (entity.__vert.__id && entity.__frag.__id) {
+      progid = _WebGL_getProgID(entity.__vert.__id, entity.__frag.__id);
+      program = model.__cache.programs[progid];
+    }
+
+    if (!program) {
+
+      var vshader;
+      if (entity.__vert.__id) {
+        vshader = model.__cache.shaders[entity.__vert.__id];
+      } else {
+        entity.__vert.__id = _WebGL_guid++;
+      }
+
+      if (!vshader) {
+        vshader = _WebGL_doCompile(gl, entity.__vert.src, gl.VERTEX_SHADER);
+        model.__cache.shaders[entity.__vert.__id] = vshader;
+      }
+
+      var fshader;
+      if (entity.__frag.__id) {
+        fshader = model.__cache.shaders[entity.__frag.__id];
+      } else {
+        entity.__frag.__id = _WebGL_guid++;
+      }
+
+      if (!fshader) {
+        fshader = _WebGL_doCompile(gl, entity.__frag.src, gl.FRAGMENT_SHADER);
+        model.__cache.shaders[entity.__frag.__id] = fshader;
+      }
+
+      program = _WebGL_doLink(gl, vshader, fshader);
+      progid = _WebGL_getProgID(entity.__vert.__id, entity.__frag.__id);
+      model.__cache.programs[progid] = program;
+
+    }
+
+    gl.useProgram(program);
+
+    progid = progid || _WebGL_getProgID(entity.__vert.__id, entity.__frag.__id);
+    var setters = model.__cache.uniformSetters[progid];
+    if (!setters) {
+      setters = _WebGL_createUniformSetters(gl, model, program);
+      model.__cache.uniformSetters[progid] = setters;
+    }
+
+    _WebGL_setUniforms(setters, entity.__uniforms);
+
+    var buffer = model.__cache.buffers[entity.__mesh.__id];
+
+    if (!buffer) {
+      buffer = _WebGL_doBindSetup(gl, entity.__mesh);
+      model.__cache.buffers[entity.__mesh.__id] = buffer;
+    }
+
+    gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, buffer.indexBuffer);
+
+    var numAttributes = gl.getProgramParameter(program, gl.ACTIVE_ATTRIBUTES);
+
+    for (var i = 0; i < numAttributes; i++) {
+      var attribute = gl.getActiveAttrib(program, i);
+
+      var attribLocation = gl.getAttribLocation(program, attribute.name);
+      gl.enableVertexAttribArray(attribLocation);
+
+      if (buffer.buffers[attribute.name] === undefined) {
+        buffer.buffers[attribute.name] = _WebGL_doBindAttribute(gl, attribute, entity.__mesh);
+      }
+      var attributeBuffer = buffer.buffers[attribute.name];
+      var attributeInfo = _WebGL_getAttributeInfo(gl, attribute.type);
+
+      gl.bindBuffer(gl.ARRAY_BUFFER, attributeBuffer);
+      gl.vertexAttribPointer(attribLocation, attributeInfo.size, attributeInfo.baseType, false, 0, 0);
+    }
+
+    _WebGL_listEach(function (setting) {
+      _WebGL_applySetting(gl, setting);
+    }, entity.__settings);
+
+    gl.drawElements(entity.__mesh.a.mode, buffer.numIndices, gl.UNSIGNED_SHORT, 0);
+
+    _WebGL_listEach(function (setting) {
+      _WebGL_revertSetting(gl, setting);
+    }, entity.__settings);
+
+  }
+
+  _WebGL_listEach(drawEntity, model.__entities);
+  return domNode;
+});
+
+function _WebGL_createUniformSetters(gl, model, program) {
+  var textureCounter = 0;
+  function createUniformSetter(program, uniform) {
+    var uniformLocation = gl.getUniformLocation(program, uniform.name);
+    switch (uniform.type) {
+      case gl.INT:
+        return function (value) {
+          gl.uniform1i(uniformLocation, value);
+        };
+      case gl.FLOAT:
+        return function (value) {
+          gl.uniform1f(uniformLocation, value);
+        };
+      case gl.FLOAT_VEC2:
+        return function (value) {
+          gl.uniform2fv(uniformLocation, new Float32Array(value));
+        };
+      case gl.FLOAT_VEC3:
+        return function (value) {
+          gl.uniform3fv(uniformLocation, new Float32Array(value));
+        };
+      case gl.FLOAT_VEC4:
+        return function (value) {
+          gl.uniform4fv(uniformLocation, new Float32Array(value));
+        };
+      case gl.FLOAT_MAT4:
+        return function (value) {
+          gl.uniformMatrix4fv(uniformLocation, false, new Float32Array(value));
+        };
+      case gl.SAMPLER_2D:
+        var currentTexture = textureCounter++;
+        return function (texture) {
+          gl.activeTexture(gl.TEXTURE0 + currentTexture);
+          var tex = model.__cache.textures[texture.__id];
+          if (!tex) {
+            _WebGL_log('Created texture');
+            tex = texture.__createTexture(gl);
+            model.__cache.textures[texture.__id] = tex;
+          }
+          gl.bindTexture(gl.TEXTURE_2D, tex);
+          gl.uniform1i(uniformLocation, currentTexture);
+        };
+      case gl.BOOL:
+        return function (value) {
+          gl.uniform1i(uniformLocation, value);
+        };
+      default:
+        _WebGL_log('Unsupported uniform type: ' + uniform.type);
+        return function () {};
+    }
+  }
+
+  var uniformSetters = {};
+  var numUniforms = gl.getProgramParameter(program, gl.ACTIVE_UNIFORMS);
+  for (var i = 0; i < numUniforms; i++) {
+    var uniform = gl.getActiveUniform(program, i);
+    uniformSetters[uniform.name] = createUniformSetter(program, uniform);
+  }
+
+  return uniformSetters;
+}
+
+function _WebGL_setUniforms(setters, values) {
+  Object.keys(values).forEach(function (name) {
+    var setter = setters[name];
+    if (setter) {
+      setter(values[name]);
+    }
+  });
+}
+
+// VIRTUAL-DOM WIDGET
+
+// eslint-disable-next-line no-unused-vars
+var _WebGL_toHtml = F3(function (options, factList, entities) {
+  return _VirtualDom_custom(
+    factList,
+    {
+      __entities: entities,
+      __cache: {},
+      __options: options
+    },
+    _WebGL_render,
+    _WebGL_diff
+  );
+});
+
+/**
+ *  Creates canvas and schedules initial _WebGL_drawGL
+ *  @param {Object} model
+ *  @param {Object} model.__cache that may contain the following properties:
+           gl, shaders, programs, uniformSetters, buffers, textures
+ *  @param {List<Option>} model.__options list of options coming from Elm
+ *  @param {List<Entity>} model.__entities list of entities coming from Elm
+ *  @return {HTMLElement} <canvas> if WebGL is supported, otherwise a <div>
+ */
+function _WebGL_render(model) {
+
+  var contextAttributes = {
+    alpha: false,
+    depth: false,
+    stencil: false,
+    antialias: false,
+    premultipliedAlpha: false
+  };
+  var sceneSettings = [];
+
+  _WebGL_listEach(function (option) {
+    var s1 = option.a;
+    // TODO: how to check $?
+    switch (option.$) {
+      case 'Alpha':
+        contextAttributes.alpha = true;
+        contextAttributes.premultipliedAlpha = s1;
+        break;
+      case 'Antialias':
+        contextAttributes.antialias = true;
+        break;
+      case 'Depth':
+        contextAttributes.depth = true;
+        sceneSettings.push(function (gl) {
+          gl.clearDepth(s1);
+        });
+        break;
+      case 'ClearColor':
+        sceneSettings.push(function (gl) {
+          gl.clearColor(option.a, option.b, option.c, option.d);
+        });
+        break;
+      case 'Stencil':
+        contextAttributes.stencil = true;
+        sceneSettings.push(function (gl) {
+          gl.clearStencil(s1);
+        });
+        break;
+    }
+  }, model.__options);
+
+  _WebGL_log('Render canvas');
+  var canvas = _VirtualDom_doc.createElement('canvas');
+  var gl = canvas.getContext && (
+    canvas.getContext('webgl', contextAttributes) ||
+    canvas.getContext('experimental-webgl', contextAttributes)
+  );
+
+  if (gl) {
+    sceneSettings.forEach(function (sceneSetting) {
+      sceneSetting(gl);
+    });
+  } else {
+    canvas = _VirtualDom_doc.createElement('div');
+    canvas.innerHTML = '<a href="http://get.webgl.org/">Enable WebGL</a> to see this content!';
+  }
+
+  model.__cache.gl = gl;
+  model.__cache.shaders = [];
+  model.__cache.programs = {};
+  model.__cache.uniformSetters = {};
+  model.__cache.buffers = [];
+  model.__cache.textures = [];
+
+  // Render for the first time.
+  // This has to be done in animation frame,
+  // because the canvas is not in the DOM yet
+  _WebGL_rAF(function () {
+    A2(_WebGL_drawGL, model, canvas);
+  });
+
+  return canvas;
+}
+
+function _WebGL_diff(oldModel, newModel) {
+  newModel.model.__cache = oldModel.model.__cache;
+  return _WebGL_drawGL(newModel);
+}
